requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
pydantic==2.11.7
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
scipy==1.14.0
python-dotenv==1.0.1
appwrite==11.0.0
openpyxl==3.1.5
pyarrow
sentence-transformers
